requirements.txt
# local package
-e .

pre-commit

# external requirements
click
Sphinx
coverage
awscli>=1.27.94
flake8
python-dotenv>=0.5.1

lightning
wandb
transformers
peft
huggingface-hub
jupyter
openai
langchain
chromadb
pinecone-client
unstructured
tiktoken
kor
# memorizing-transformers-pytorch

pdfplumber
PyPDF2
diffusers
whisper

fastapi
uvicorn
python-multipart
boto3
bitsandbytes
google-cloud-secret-manager
datasets
accelerate
loralib
pymupdf
